requirements.txt
gitdb==4.0.11
GitPython==3.1.41
grpcio==1.62.0
grpcio-tools==1.62.0
numpy==1.26.4
paho-mqtt==2.0.0
protobuf==4.25.3
six==1.16.0
smmap==5.0.1
